Update README.md
# Floralclassification
Abstract of project
There are hundreds of kinds of trees in the natural ecosystem, and it can be very difficult to distinguish
between them. Botanists and those who study plants however, are able to identify the type of tree at a
glance by using the characteristics of the leaf. Machine learning is used to automatically classify leaf
types. Studied extensively in 2012, this is a rapidly growing field based on deep learning. Deep learning
is itself a self-learning technique used on large amounts of data, and recent developments in hardware
and big data have made this technique more practical. We propose a method to classify leaves using the
CNN model, which is often used when applying deep learning to image processing.
First challenges introduced by illumination changes and deblurring are eliminated with some
preprocessing steps. Following the preprocessing step, Convolutional Neural Network architecture is
employed to extract the features of images. The construction of the CNN architecture and the depth of
CNN are crucial points that should be emphasized since they affect the recognition capability of the
architecture of neural networks. Thus, Leaf features are extracted and orthogonalized into principal
variables which consist the input of the CNN. The images are trained number of times to minimize the
loss parameter and optimize the results to get maximum accuracy leaves to classify 6 kinds of plants
with an accuracy nearly equal to 90%. Compared with other approaches, our algorithm is an accurate
artificial intelligence approach which is fast in execution and easy in implementation.